Made Contact Us form more user-friendly with JavaScript
If JavaScript is enabled, teh form will stay on the same page and just show a visual indicator that the form is submitting instead of redirecting to another page.

If JavaScript is disabled, the form will still function as normal but will redirect to another page showing the confirmation instead.