Changed students page and upgraded Gulp translator
Students page has been changed to be a little more intuitive in linking to other things instead of just being a big red and yellow chart taking up the whole page. More changes to come.

The Gulp translator has been completely redone; it now uses Google's Cloud Translation API instead of the old hacky way that constantly broke if you used it too much. To translate, you do need a credentials file on your computer, so it's not open for everyone unless you have your own project set up with Translation on it (which is free for a smaller website with not that much text like this one). Besides switching to the Cloud Translation API, the entire way pages are translated to begin with has been majorly improved. The code is a lot easier to work with now (it no longer uses Gulp streams, which was kind of a hassle to work around). Only files that have been changed since the last commit will be looked at, and each string of text that gets translated is done entirely asynchronously, even for collections, so every single thing gets translated all at the same time. (It only took 1.18s for me to generate all the Spanish files in this commit).